+++ src/server/rpc-lua/rpc-lua.cpp
@@ -457,13 +457,13 @@ static void init_callbacks() {
   callbacks.negint64 = [](void* self, uint64_t value) {
     static_cast<RpcPacketBuilder*>(self)->handleInteger(-1 - static_cast<int64_t>(value));
   };
-  callbacks.byte_string = [](void* self, const cbor_data data, size_t len) {
+  callbacks.byte_string = [](void* self, const cbor_data data, uint64_t len) {
     static_cast<RpcPacketBuilder*>(self)->handleBytes(data, len);
   };
-  callbacks.array_start = [](void* self, size_t size) {
+  callbacks.array_start = [](void* self, uint64_t size) {
     static_cast<RpcPacketBuilder*>(self)->startArray(size);
   };
-  callbacks.map_start = [](void* self, size_t size) {
+  callbacks.map_start = [](void* self, uint64_t size) {
     static_cast<RpcPacketBuilder*>(self)->startMap(size);
   };
   callbacks.boolean = [](void* self, bool value) {
